feat: integrate address format with SuiAddress
