+++ packages/axelar-wasm-std/Cargo.toml
@@ -45,6 +45,7 @@ report = { workspace = true }
 schemars = "0.8.10"
 serde = { version = "1.0.145", default-features = false, features = ["derive"] }
 serde_json = "1.0.89"
+sui-types = { workspace = true }
 sha3 = { workspace = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
