+++ packages/axelar-wasm-std/src/address.rs
@@ -2,6 +2,8 @@ use alloy_primitives::Address;
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, Api};
 use error_stack::{Result, ResultExt};
+use std::str::FromStr;
+use sui_types::SuiAddress;
 
 #[derive(thiserror::Error)]
 #[cw_serde]
@@ -13,13 +15,20 @@ pub enum Error {
 #[cw_serde]
 pub enum AddressFormat {
     Eip55,
+    Sui,
 }
 
 pub fn validate_address(address: &str, format: &AddressFormat) -> Result<(), Error> {
     match format {
-        AddressFormat::Eip55 => Address::parse_checksummed(address, None)
-            .change_context(Error::InvalidAddress(address.to_string()))?,
-    };
+        AddressFormat::Eip55 => {
+            Address::parse_checksummed(address, None)
+                .change_context(Error::InvalidAddress(address.to_string()))?;
+        }
+        AddressFormat::Sui => {
+            SuiAddress::from_str(address)
+                .change_context(Error::InvalidAddress(address.to_string()))?;
+        }
+    }
 
     Ok(())
 }
@@ -36,7 +45,7 @@ mod tests {
     use crate::{address, err_contains};
 
     #[test]
-    fn validate_address() {
+    fn validate_eip55_address() {
         let addr = "0x95222290DD7278Aa3Ddd389Cc1E1d165CC4BAfe5";
 
         assert!(address::validate_address(addr, &address::AddressFormat::Eip55).is_ok());
@@ -49,6 +58,22 @@ mod tests {
 
         let upper_case = addr.to_uppercase();
         assert!(address::validate_address(&upper_case, &address::AddressFormat::Eip55).is_err());
+    }
+
+    #[test]
+    fn validate_sui_address() {
+        let addr = "0x8cc8d18733a4bf98de8f861d356e2191918733e3afff29f327a01b5ba2997a4d";
+
+        assert!(address::validate_address(addr, &address::AddressFormat::Sui).is_ok());
+
+        let without_prefix = addr.strip_prefix("0x").unwrap();
+        assert!(address::validate_address(without_prefix, &address::AddressFormat::Sui).is_err());
+
+        let upper_case = addr.to_uppercase();
+        assert!(address::validate_address(&upper_case, &address::AddressFormat::Sui).is_err());
+
+        let invalid_length = format!("{}5f", addr);
+        assert!(address::validate_address(&invalid_length, &address::AddressFormat::Sui).is_err());
     }
 
     #[test]
