+++ packages/sui-types/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "sui-types"
+version = "1.0.0"
+edition = { workspace = true }
+rust-version = { workspace = true }
+
+[dependencies]
+error-stack = { workspace = true }
+hex = { workspace = true }
+serde = { workspace = true }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+rand = "0.8.5"
+
+[lints]
+workspace = true
